+++ packages/adminui-pub/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+ - set default width for multifilter combobox
 
 ## [5.14.5] - 2021-01-26
 ### Added